requirements.txt
astroid==3.0.0
blinker==1.6.2
click==8.1.7
dill==0.3.7
Flask==2.3.3
Flask-MySQLdb==1.0.1
isort==5.12.0
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
MarkupSafe==2.1.3
mccabe==0.7.0
mysqlclient==2.2.0
parso==0.8.3
platformdirs==3.11.0
PyJWT==2.8.0
pylint==3.0.0
tomli==2.0.1
tomlkit==0.12.1
typing_extensions==4.8.0
Werkzeug==3.0.0
